chore: restructure examples to match documentation
