+++ examples/web-wagmi/src/discovery/DiscoveryPage.tsx
@@ -6,6 +6,26 @@ const discoveryHooks = [
     description: `View the feed for a profile.`,
     path: '/discovery/useFeed',
   },
+  {
+    label: 'useExploreProfiles',
+    description: `Explore different profiles.`,
+    path: '/discovery/useExploreProfiles',
+  },
+  {
+    label: 'useExplorePublications',
+    description: `Explore publications that match certain criteria.`,
+    path: '/discovery/useExplorePublications',
+  },
+  {
+    label: 'useSearchProfiles',
+    description: `Fetch profiles that match a search query`,
+    path: '/discovery/useSearchProfiles',
+  },
+  {
+    label: 'useSearchPublications',
+    description: `Fetch publications that match a search query`,
+    path: '/discovery/useSearchPublications',
+  },
 ];
 
 export function DiscoveryPage() {
